==> 3. Combined filtering.ipynb <==
Last updated: July 5, 2023

Last run: July 5, 2023

**Additional filtering based on data (as described in the published study)**

## Unreviewed science in the news: The evolution of preprint media coverage from 2014-2021


Juan Pablo Alperin

**Related Publication:**
Fleerackers, A., Shores, K., Chtena, N. & Alperin, J.P. (2023). Unreviewed science in the news: The evolution of preprint media coverage from 2014-2021. *bioarxiv*. 

**Related Dataset:**
Alperin, Juan Pablo; Fleerackers; Shores, 2023, "Data for: Unreviewed science in the news", https://doi.org/10.7910/DVN/ZHQUFD, *Harvard Dataverse*

import pandas as pd

## Preprint filtering

df = pd.read_csv('data/top outlets/preprints_mention_details_with_dates.csv')

df['posted_on'] = pd.to_datetime(df.posted_on)

df['best_published_pub_date'] = pd.to_datetime(df['best_published_pub_date'])
df['best_preprint_pub_date'] = pd.to_datetime(df['best_preprint_pub_date'])

preprints_original = df.copy()

N = df.shape[0]
print("{:,}".format(N))

# Remove everything before 2013 that appeared from problematic dates
df = df[df.best_preprint_pub_date >= '2013']

print("{:,}".format(N-df.shape[0]))
N=df.shape[0]

# filter those that appear to actually be postprints, not preprints
df = df[(df.best_published_pub_date.isnull()) | (df.best_published_pub_date > df.best_preprint_pub_date)]

print("{:,}".format(N - df.shape[0]))
N=df.shape[0]

# remaining
print("{:,}".format(df.shape[0]))

# date data too uncertain to decide if preprint or postprint (dates too closer together)
# 7 days 

df = df[(df.best_published_pub_date.isnull()) | (abs((df.best_published_pub_date - df.best_preprint_pub_date).dt.days) > 7)]

print("{:,}".format(N - df.shape[0]))
N=df.shape[0]

# The news media mentioned happened more than 5 days before the preprint was published

df = df[df.days_since_preprint >= -5]

print("{:,}".format(N - df.shape[0]))
N=df.shape[0]

# mention appears to be to the postprint, not preprint

df = df[(df.days_since_publication.isnull()) | (df.days_since_publication < -5)]

print("{:,}".format(N - df.shape[0]))
N=df.shape[0]

# We have the same news story for the same DOI/arxiv_id more than once
df = df.drop_duplicates(subset=['doi', 'arxiv_id', 'outlet', 'moreover_url'])

print("{:,}".format(N - df.shape[0]))
N=df.shape[0]

df.to_csv('data/top outlets/preprints_mentions_final.csv', index=False)

preprints = df.copy()
del df

## Summary of preprint filtering

N_mentions = preprints.shape[0]
N_mentions_original = preprints_original.shape[0]

N_preprints = preprints.altmetric_id.unique().shape[0]
N_preprints_original = preprints_original.altmetric_id.unique().shape[0]

N_stories =  preprints[['outlet', 'moreover_url']].drop_duplicates().shape[0]
N_stories_original =  preprints_original[['outlet', 'moreover_url']].drop_duplicates().shape[0]

N_outlets = preprints.outlet.unique().shape[0]

print("{:,} mentions ({:.1f}%)".format(N_mentions, 100*N_mentions/N_mentions_original))
print("{:,} preprints ({:.1f}%)".format(N_preprints, 100*N_preprints/N_preprints_original))
print("{:,} stories ({:.1f}%)".format(N_stories, 100*N_stories/N_stories_original))
print()
print("In total, filtering led to the exclusion of {:,} mentions ({:.1f}% of original dataset)".format(
    N_mentions_original-N_mentions, 100*(N_mentions_original-N_mentions)/N_mentions_original))
print()
print("Our final preprint sample comprised {:,} mentions of {:,} preprints in {:,} stories published by the 99 outlets in our sample (a preprint could be covered in several stories and a single story could mention multiple preprints)".format( 
     N_mentions, N_preprints, N_stories))



## WoS filtering

input_file = 'data/top outlets/wos_with_doi_mention_details.csv'
# columns = ['Altmetric_ID', 'Order', 'DOI', 'Arxiv_ID', 'First_Seen_On', 'PubDate', 'wos_year', 'wos_title', 'Author_Name', 'Author_Url', 'Posted_On', 'Title', 'moreover_url']

df = pd.read_csv(input_file)
df.columns = [x.lower() for x in df.columns]
df.rename(columns = {'author_name': 'outlet', 'author_url': 'outlet_url'}, inplace=True)
df['posted_on'] = pd.to_datetime(df.posted_on)
df['posted_on_year'] = df.posted_on.map(lambda x: x.year)

print("{:,}".format(df.shape[0]))

articles_original = df.copy()

# Remove everything before 2013, to match what we did with preprints because of problematic dates
N = df.shape[0]
df = df[df.wos_year >= 2013]

print("{:,}".format(N - df.shape[0]))
N=df.shape[0]

# Merging WoS with Preprint data to find duplicates, etc.
merged = df.set_index(['altmetric_id', 'order']).merge(preprints[['altmetric_id', 'order', 'posted_on']].set_index(['altmetric_id', 'order']), 
         left_index=True, right_index=True, 
         how='outer', 
        suffixes=['', '_p'],
         indicator='source'
        )

print("Removing {:,} WoS mentions that are in the preprint mentions folder".format(merged[merged.source == 'both'].shape[0]))

df = merged[merged.source == 'left_only']
del df['source']
del df['posted_on_p']

# We have the same news story for the same DOI more than once
df = df.drop_duplicates(subset=['doi', 'outlet', 'moreover_url'])

print("{:,}".format(N - df.shape[0]))
N=df.shape[0]

print("{:,}".format(df.shape[0]))
df.to_csv('data/top outlets/wos_mentions_final.csv')

articles = df.copy()
del df

## Summary of WoS filtering

N_mentions = articles.shape[0]
N_mentions_original = articles_original.shape[0]

N_articles = articles.doi.unique().shape[0]
N_articles_original = articles_original.doi.unique().shape[0]

N_stories =  articles[['outlet', 'moreover_url']].drop_duplicates().shape[0]
N_stories_original =  articles_original[['outlet', 'moreover_url']].drop_duplicates().shape[0]

N_outlets = articles.outlet.unique().shape[0]

print("{:,} mentions ({:.1f}%)".format(N_mentions, 100*N_mentions/N_mentions_original))
print("{:,} preprints ({:.1f}%)".format(N_preprints, 100*N_preprints/N_preprints_original))
print("{:,} stories ({:.1f}%)".format(N_stories, 100*N_stories/N_stories_original))
print()
print("In total, filtering led to the exclusion of {:,} mentions ({:.1f}% of original dataset)".format(
    N_mentions_original-N_mentions, 100*(N_mentions_original-N_mentions)/N_mentions_original))
print()
print("The final published research sample comprised {:,} mentions of {:,} distinct research outputs across {:,} stories published by the {:,} outlets in our sample (a research output could be covered in several stories and a single story could mention multiple outputs)".format( 
     N_mentions, N_articles, N_stories, N_outlets))



"A total of {:,} mentions of {:,} publications across {:,} news stories.".format(articles.shape[0], articles.doi.unique().shape[0], articles[['outlet', 'moreover_url']].drop_duplicates().shape[0])
