==> 5. Merge datasets.ipynb <==
Last updated: July 5, 2023

Last run: July 5, 2023

**Additional filtering based on data (as described in the published study)**

## Unreviewed science in the news: The evolution of preprint media coverage from 2014-2021


Juan Pablo Alperin

**Related Publication:**
Fleerackers, A., Shores, K., Chtena, N. & Alperin, J.P. (2023). Unreviewed science in the news: The evolution of preprint media coverage from 2014-2021. *bioarxiv*. 

**Related Dataset:**
Alperin, J. P., Fleerackers, A., & Shores, K. (2023). Data for: Unreviewed science in the news [Data set]. Harvard Dataverse. https://doi.org/10.7910/DVN/ZHQUFD

import pandas as pd

preprints = pd.read_csv('data/top outlets/preprint_mentions_w_coding.csv')
wos = pd.read_csv('data/top outlets/wos_mentions_w_coding.csv')

columns = ['altmetric_id', 'doi', 'arxiv_id', 'posted_on', 'pub_year', 'is_covid', 'outlet_name', 'outlet_type', 'outlet_topic', 'outlet_location', 'research_venue']

df = wos[['altmetric_id', 'doi', 'arxiv_id', 'posted_on', 'wos_year', 'covid_title', 'outlet', 'Type', 'Topic', 'Location']]
df['research_venue'] = 'wos'

df.columns = columns

df2 = preprints[['altmetric_id', 'doi', 'arxiv_id', 'posted_on', 'best_preprint_pub_date', 'covid_title', 'outlet', 'Type', 'Topic', 'Location', 'server']]
df2.columns = columns


final = df.append(df2)

final['posted_on_year'] = final.posted_on.map(lambda x: int(str(x)[0:4]))
final['pub_year'] = final.pub_year.map(lambda x: int(str(x)[0:4]))
final['is_covid'] = final.is_covid.astype(int)
final['is_preprint'] = final.research_venue.map(lambda x: int(x != 'wos'))

final['posted_on'] = pd.to_datetime(final.posted_on)
min_date = final.posted_on.min()
final['posted_on_day_number'] = final['posted_on'].map(lambda x: (x - min_date).days)
final['posted_on_year'] = final.posted_on.map(lambda x: x.year)
final['posted_on_month'] = final.posted_on.map(lambda x: x.month)
final['posted_on_day'] = final.posted_on.map(lambda x: x.day)

final['article_id'] = final.apply(lambda row: "{}_{}".format(row['altmetric_id'], ('p' if row['is_preprint'] else 'w')), axis=1)

final.sort_values(['posted_on', 'article_id'], inplace=True)

final.reset_index(drop=True, inplace=True)
final.index.name = 'mention_id'

final.to_csv('data/research_mentions_all_final.csv')
